backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
requests==2.28.1
python-dotenv==0.20.0
loguru==0.6.0
apscheduler==3.10.1
pytz==2022.7.1
beautifulsoup4==4.11.1 